requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
urllib3==2.0.7
requests==2.31.0 